src/test/javascript/spec/app/entities/device-model/device-model.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { DeviceModelService } from 'app/entities/device-model/device-model.service';
import { IDeviceModel, DeviceModel } from 'app/shared/model/device-model.model';
import { DeviceType } from 'app/shared/model/enumerations/device-type.model';

describe('Service Tests', () => {
  describe('DeviceModel Service', () => {
    let injector: TestBed;
    let service: DeviceModelService;
    let httpMock: HttpTestingController;
    let elemDefault: IDeviceModel;
    let expectedResult: IDeviceModel | IDeviceModel[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(DeviceModelService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new DeviceModel(0, 'AAAAAAA', false, 0, 'image/png', 'AAAAAAA', 'image/png', 'AAAAAAA', DeviceType.IPPHONE);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a DeviceModel', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new DeviceModel()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a DeviceModel', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            isConfigurable: true,
            linesCount: 1,
            configTemplate: 'BBBBBB',
            firmwareFile: 'BBBBBB',
            deviceType: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of DeviceModel', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            isConfigurable: true,
            linesCount: 1,
            configTemplate: 'BBBBBB',
            firmwareFile: 'BBBBBB',
            deviceType: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a DeviceModel', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
